Function spotify_party_api::utils::requests_to_api::refresh_put_request_empty::refresh_and_send_empty_put_request

pub async fn refresh_and_send_empty_put_request(
    req: HttpRequest,
    auth_header: &str,
    url: &str
) -> (bool, HttpResponse)

This function is responsible for refreshing the access token and then sending an empty PUT request.

Arguments

req - An instance of HttpRequest which contains the request data.

auth_header - A reference to a string representing the authorization header value.

url - A reference to a string representing the URL to send the PUT request to.

Returns

A tuple containing a boolean and an HttpResponse. The boolean indicates whether the token was refreshed successfully (true) or not (false). The HttpResponse represents the response from the server after sending the PUT request. If the token was refreshed successfully and the PUT request was successful, the HttpResponse will be NoContent (204). If the token could not be refreshed or the PUT request failed, the HttpResponse will be BadRequest (400).